livekit/requirements.txt
# LiveKit Voice Agent Requirements

# Core LiveKit dependencies
livekit-agents>=1.2.11
livekit-plugins-openai>=0.8.0
livekit-plugins-silero>=0.8.0
livekit-plugins-deepgram>=0.8.0
livekit-agents[cartesia]~=1.2
livekit-plugins-cartesia>=1.2.17

# OpenAI integration
openai>=1.0.0

# Database
motor>=3.3.0
dnspython>=2.3.0

# HTTP client
aiohttp>=3.8.0
httpx>=0.28.0

# Environment management
python-dotenv>=1.0.0

# Date/time handling
python-dateutil>=2.8.0
tzdata>=2023.3

# Logging and monitoring
structlog>=23.0.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Development tools (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0